make : delete the '-Wno-sign-compare' from both CFLAGS and CXXFLAGS
Fixes #350;
Compile cmd is:git clean -dxf && ./autogen.sh --prefix=/home/wdp/yami_bin//libyami
                 && make -j8 && make install
variable TEST_CONFIGURE: export TEST_CONFIGURE="--enable-silent-rules --enable-debug
         --enable-tests --enable-tests-gles --enable-dmabuf --enable-v4l2 --enable-capi
         --enable-x11 --enable-avformat --enable-baytrail --enable-v4l2-ops --enable-md5
         --enable-vp8dec --enable-vp9dec --enable-jpegdec --enable-h264dec --enable-h265dec
         --enable-fakedec --enable-h264enc --enable-jpegenc --enable-vp8enc --enable-h265enc
         --enable-docs --enable-dependency-tracking --disable-dependency-tracking
         --disable-libtool-lock"
variable TEST_CONFIGURE excludes '--enable-getmv' and  '--enable-v4l2-glx'

+++ interface/VideoDecoderDefs.h
@@ -103,7 +103,7 @@ typedef enum {
 
 typedef struct {
     uint8_t *data;
-    int32_t size;
+    size_t size;
     int64_t timeStamp;
     uint32_t flag;
     VideoExtensionBuffer *ext;
@@ -157,8 +157,8 @@ typedef struct SurfaceBuffer{
 typedef struct {
     bool valid;                 // indicates whether format info is valid. MimeType is always valid.
     char *mimeType;
-    int32_t width;
-    int32_t height;
+    uint32_t width;
+    uint32_t height;
     int32_t surfaceWidth;
     int32_t surfaceHeight;
     int32_t surfaceNumber;
